Improve performance in cleanup()
We remove directories by calling an external program.  In the case of
purging many directories, we will need to execute the external process
as many times.  Since the cleanup() should execute as fast as possible,
this might be a bottleneck.

Instead, let find(1) start the external process, but let it first
collect all necessary paths and then pass ALL of them to a single
invocation of rm(1).

Another speedup in this category would be the merging of the previous 2
invocations of rm(1) into one, although such a thing would slightly hurt
readability and maintainability of the function.